Add Via + Vial compat
Change a bit the outline of the readme to give more space to the practical information and put the nice pictures and renders at the end of the document.

Update Readme with last set of pictures made on the acrylic case.

Add Via + Vial compat

TBD info.json in firmware folder